Fix tests

+++ src/test/scala/is/hail/methods/AnnotateGlobalSuite.scala
@@ -24,7 +24,7 @@ class AnnotateGlobalSuite extends SparkSuite {
     val qSingleton = vds.querySA("sa.qc.nSingleton")._2
 
     val sCount = vds.sampleAnnotations.count(sa =>
-      qSingleton(sa).asInstanceOf[Int] > 2)
+      qSingleton(sa).asInstanceOf[Long] > 2)
 
     assert(singStats == sCount)
